[Fix] 포트폴리오 타입 수정
