+++ src/types/portfolio.ts
@@ -1,0 +1,20 @@
+export interface PortfolioItem {
+  id: string
+  userId: string
+  title: string
+  content: string
+  fileList: string[]
+  interest: string
+  career: string
+  isBookmarked: boolean
+  linkUrl: string
+  createdAt: string
+  viewCount: number
+  likeCount: number
+  techStack: string[]
+  email: string
+  name: string
+  phone: string
+  profileImage: string
+  birthDate: string
+}
